feedback_system/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
sqlalchemy
aiosqlite
pytest
pytest-asyncio
httpx
greenlet
torch
transformers
sentencepiece
rich
